Fixing try scenario'
