+++ AzureFunctions.AngularClient/src/app/app.module.ts
@@ -1,3 +1,4 @@
+import { SlotsService } from 'app/shared/services/slots.service';
 import { CreateAppComponent } from './site/create-app/create-app.component';
 import { ClickToEditComponent } from './controls/click-to-edit/click-to-edit.component';
 import { AiTryService } from './shared/services/ai-try.service';
@@ -6,7 +7,6 @@ import { TblThComponent } from './controls/tbl/tbl-th/tbl-th.component';
 import { TblComponent } from './controls/tbl/tbl.component';
 import { GlobalErrorHandler } from './shared/GlobalErrorHandler';
 import { ErrorHandler } from '@angular/core';
-import { ArmTryService } from './shared/services/arm-try.service';
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule, APP_INITIALIZER } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
@@ -20,8 +20,9 @@ import { FunctionsService } from './shared/services/functions.service';
 import { UserService } from './shared/services/user.service';
 import { PortalService } from './shared/services/portal.service';
 import { BroadcastService } from './shared/services/broadcast.service';
-import { FunctionMonitorService } from './shared/services/function-monitor.service';
-import { ArmService } from './shared/services/arm.service';
+
+import { FunctionMonitorService } from './shared/services/function-monitor.service'
+import { ArmService, ArmServiceBase } from './shared/services/arm.service';
 import { CacheService } from './shared/services/cache.service';
 import { AuthzService } from './shared/services/authz.service';
 import { LocalStorageService } from './shared/services/local-storage.service';
@@ -104,148 +105,136 @@ import { SiteConfigComponent } from './site/site-config/site-config.component';
 import { CommandBarComponent } from './controls/command-bar/command-bar.component';
 import { CommandComponent } from './controls/command-bar/command/command.component';
 import { SlotsListComponent } from './slots-list/slots-list.component';
-import { SlotsService } from './shared/services/slots.service';
 import { SlotNewComponent } from './slot-new/slot-new.component';
 
 export function ConfigLoader(config: ConfigService) {
   return () => config.loadConfig();
 }
 
-export function ArmServiceLoader() {
-  return window.location.pathname.toLowerCase() === '/try' ? ArmTryService : ArmService;
-}
-export function AiServiceLoader() {
-  return window.location.pathname.toLowerCase() === '/try' ? AiTryService : AiService;
-}
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    GettingStartedComponent,
-    BusyStateComponent,
-    TryNowBusyStateComponent,
-    TopBarComponent,
-    DropDownComponent,
-    TryNowComponent,
-    FunctionEditComponent,
-    TrialExpiredComponent,
-    FunctionNewComponent,
-    FunctionQuickstartComponent,
-    TutorialComponent,
-    SourceControlComponent,
-    FunctionDevComponent,
-    BindingComponent,
-    TooltipContentComponent,
-    TooltipDirective,
-    ErrorListComponent,
-    TemplatePickerComponent,
-    PopOverComponent,
-    BindingInputComponent,
-    BindingDesignerComponent,
-    SecretsBoxContainerComponent,
-    SecretsBoxInputDirective,
-    AggregateBlockComponent,
-    CopyPreComponent,
-    FileExplorerComponent,
-    FunctionIntegrateV2Component,
-    FunctionIntegrateComponent,
-    FunctionKeysComponent,
-    FunctionManageComponent,
-    FunctionMonitorComponent,
-    LogStreamingComponent,
-    RadioSelectorComponent,
-    RunHttpComponent,
-    TableFunctionMonitorComponent,
-    TryLandingComponent,
-    AggregateBlockPipe,
-    MonacoEditorDirective,
-    TableFunctionMonitorPipe,
-    MainComponent,
-    SideNavComponent,
-    TreeViewComponent,
-    SiteDashboardComponent,
-    TabsComponent,
-    TabComponent,
-    BreadcrumbsComponent,
-    SiteSummaryComponent,
-    SiteEnabledFeaturesComponent,
-    SiteManageComponent,
-    FeatureGroupComponent,
-    DeploymentSourceComponent,
-    DeploymentSourceSetupComponent,
-    MultiDropDownComponent,
-    TopRightMenuComponent,
-    AppsListComponent,
-    FunctionRuntimeComponent,
-    ApiDetailsComponent,
-    ApiNewComponent,
-    FunctionsListComponent,
-    ProxiesListComponent,
-    SlotsListComponent,
-    SwaggerDefinitionComponent,
-    SwaggerFrameDirective,
-    DisabledDashboardComponent,
-    CreateFunctionWrapperComponent,
-    TblComponent,
-    TblThComponent,
-    FnWriteAccessDirective,
-    EditModeWarningComponent,
-    TextboxComponent,
-    SiteConfigComponent,
-    ClickToEditComponent,
-    CommandBarComponent,
-    CommandComponent,
-    CreateAppComponent,
-    SlotsListComponent,
-    SlotNewComponent
-  ],
-  imports: [
-    FormsModule,
-    ReactiveFormsModule,
-    BrowserModule,
-    FormsModule,
-    HttpModule,
-    TranslateModule.forRoot(),
-    FileUploadModule
-  ],
-  providers: [
-    ConfigService,
-    {
-      provide: APP_INITIALIZER,
-      useFactory: ConfigLoader,
-      deps: [ConfigService],
-      multi: true
-    },
-    FunctionsService,
-    UserService,
-    LanguageService,
-    PortalService,
-    BroadcastService,
-    FunctionMonitorService,
-    //   ArmService,
-    {
-      provide: ArmService,
-      // useClass: window.location.pathname.toLowerCase() === '/try' ? ArmTryService : ArmService
-      useClass: ArmServiceLoader()
-    },
-    CacheService,
-    SlotsService,
-    AuthzService,
-    LocalStorageService,
-    TelemetryService,
-    UtilitiesService,
-    BackgroundTasksService,
-    GlobalStateService,
-    {
-      provide: AiService,
-      // useClass: window.location.pathname.toLowerCase() === '/try' ? AiTryService : AiService
-      useClass: AiServiceLoader()
-    },
-    {
-      provide: ErrorHandler,
-      useClass: GlobalErrorHandler
-    },
-  ],
-  bootstrap: [AppComponent]
+    declarations: [
+        AppComponent,
+        GettingStartedComponent,
+        BusyStateComponent,
+        TryNowBusyStateComponent,
+        TopBarComponent,
+        DropDownComponent,
+        TryNowComponent,
+        FunctionEditComponent,
+        TrialExpiredComponent,
+        FunctionNewComponent,
+        FunctionQuickstartComponent,
+        TutorialComponent,
+        SourceControlComponent,
+        FunctionDevComponent,
+        BindingComponent,
+        TooltipContentComponent,
+        TooltipDirective,
+        ErrorListComponent,
+        TemplatePickerComponent,
+        PopOverComponent,
+        BindingInputComponent,
+        BindingDesignerComponent,
+        SecretsBoxContainerComponent,
+        SecretsBoxInputDirective,
+        AggregateBlockComponent,
+        CopyPreComponent,
+        FileExplorerComponent,
+        FunctionIntegrateV2Component,
+        FunctionIntegrateComponent,
+        FunctionKeysComponent,
+        FunctionManageComponent,
+        FunctionMonitorComponent,
+        LogStreamingComponent,
+        RadioSelectorComponent,
+        RunHttpComponent,
+        TableFunctionMonitorComponent,
+        TryLandingComponent,
+        AggregateBlockPipe,
+        MonacoEditorDirective,
+        TableFunctionMonitorPipe,
+        MainComponent,
+        SideNavComponent,
+        TreeViewComponent,
+        SiteDashboardComponent,
+        TabsComponent,
+        TabComponent,
+        BreadcrumbsComponent,
+        SiteSummaryComponent,
+        SiteEnabledFeaturesComponent,
+        SiteManageComponent,
+        FeatureGroupComponent,
+        DeploymentSourceComponent,
+        DeploymentSourceSetupComponent,
+        MultiDropDownComponent,
+        TopRightMenuComponent,
+        AppsListComponent,
+        FunctionRuntimeComponent,
+        ApiDetailsComponent,
+        ApiNewComponent,
+        FunctionsListComponent,
+        ProxiesListComponent,
+        SlotsListComponent,
+        SwaggerDefinitionComponent,
+        SwaggerFrameDirective,
+        DisabledDashboardComponent,
+        CreateFunctionWrapperComponent,
+        TblComponent,
+        TblThComponent,
+        FnWriteAccessDirective,
+        EditModeWarningComponent,
+        TextboxComponent,
+        SiteConfigComponent,
+        ClickToEditComponent,
+        CommandBarComponent,
+        CommandComponent,
+        CreateAppComponent,
+        SlotsListComponent,
+        SlotNewComponent
+    ],
+    imports: [
+        FormsModule,
+        ReactiveFormsModule,
+        BrowserModule,
+        FormsModule,
+        HttpModule,
+        TranslateModule.forRoot(),
+        FileUploadModule
+    ],
+    providers: [
+        ConfigService,
+        {
+            provide: APP_INITIALIZER,
+            useFactory: ConfigLoader,
+            deps: [ConfigService],
+            multi: true
+        },
+        FunctionsService,
+        UserService,
+        LanguageService,
+        PortalService,
+        BroadcastService,
+        FunctionMonitorService,
+        SlotsService,
+        ArmServiceBase,
+        ArmService,
+        CacheService,
+        AuthzService,
+        LocalStorageService,
+        TelemetryService,
+        UtilitiesService,
+        BackgroundTasksService,
+        GlobalStateService,
+        {
+            provide: AiService,
+            useClass: window.location.pathname.toLowerCase() === '/try' ? AiTryService : AiService
+        },
+        {
+            provide: ErrorHandler,
+            useClass: GlobalErrorHandler
+        },
+    ],
+    bootstrap: [AppComponent]
 })
 export class AppModule { }
