+++ AzureFunctions.AngularClient/src/app/shared/services/arm.service.ts
@@ -1,3 +1,5 @@
+import { SiteDescriptor } from './../resourceDescriptors';
+import { FunctionApp } from './../function-app';
 import { Guid } from './../Utilities/Guid';
 import {Http, Headers, Response, Request} from '@angular/http';
 import {Injectable, EventEmitter} from '@angular/core';
@@ -30,9 +32,9 @@ import {ConfigService} from './config.service';
 // }
 
 @Injectable()
-export class ArmService {
-  public static availabilityApiVersion = '2015-01-01';
 
+export class ArmServiceBase {
+    public static availabilityApiVersion = '2015-01-01';
     public subscriptions = new ReplaySubject<Subscription[]>(1);
     public armUrl = '';
 
@@ -48,7 +50,7 @@ export class ArmService {
     private _invokeId = 100;
 
     constructor(private _http: Http,
-        private _configService: ConfigService,
+        protected _configService: ConfigService,
         private _userService: UserService,
         protected _aiService: AiService,
         private _translateService: TranslateService) {
@@ -142,5 +144,152 @@ export class ArmService {
         }
 
         return headers;
+    }
+}
+
+
+@Injectable()
+export class ArmService extends ArmServiceBase {
+
+    private _tryMode = window.location.pathname.toLowerCase() === '/try';
+    private _tryFunctionApp : FunctionApp;
+
+    private _whiteListedPrefixUrls : string[] = [
+            `${Constants.serviceHost}api`
+        ];;
+
+    constructor(http: Http,
+        configService: ConfigService,
+        userService: UserService,
+        aiService: AiService,
+        translateService: TranslateService) {
+
+        super(http, configService, userService, aiService, translateService);
+    }
+
+    public set tryFunctionApp(tryFunctionApp : FunctionApp){
+        this._tryFunctionApp = tryFunctionApp;
+        this._whiteListedPrefixUrls.push(`${tryFunctionApp.getScmUrl()}/api`);
+        this._whiteListedPrefixUrls.push(`${tryFunctionApp.getMainSiteUrl()}`);
+
+        let descriptor = new SiteDescriptor(tryFunctionApp.site.id);
+        this.subscriptions.next([{
+            subscriptionId : descriptor.subscription,
+            displayName : "Trial Subscription",
+            state: "Enabled"
+        }])
+    }
+
+    public get tryFunctionApp(){
+        return this._tryFunctionApp;
+    }
+
+    get(resourceId : string, apiVersion? : string) : Observable<Response>{
+        if(!this._tryMode){
+            return super.get(resourceId, apiVersion);
+        }
+        
+        this._aiService.trackEvent("/try/arm-get-failure", {
+            uri : resourceId
+        });
+
+        throw "[ArmTryService] - get: " + resourceId;
+    }
+
+    delete(resourceId : string, apiVersion? : string) : Observable<Response>{
+        if(!this._tryMode){
+            return super.delete(resourceId, apiVersion);
+        }
+
+        this._aiService.trackEvent("/try/arm-delete-failure", {
+            uri : resourceId
+        });
+
+        throw "[ArmTryService] - delete: " + resourceId;
+    }
+
+    put(resourceId : string, body : any, apiVersion? : string) : Observable<Response>{
+        if(!this._tryMode){
+            return super.put(resourceId, body, apiVersion);
+        }        
+        this._aiService.trackEvent("/try/arm-put-failure", {
+            uri : resourceId
+        });
+
+        throw "[ArmTryService] - put: " + resourceId;
+    }
+
+    post(resourceId : string, body : any, apiVersion? : string) : Observable<Response>{
+        if(!this._tryMode){
+            return super.post(resourceId, body, apiVersion);
+        }        
+        
+        this._aiService.trackEvent("/try/arm-post-failure", {
+            uri : resourceId
+        });
+
+        throw "[ArmTryService] - post: " + resourceId;
+    }
+
+    send(method : string, url : string, body? : any, etag? : string, headers? : Headers, invokeApi? : boolean) : Observable<Response>{
+        if(!this._tryMode){
+            return super.send(method, url, body, etag, headers, invokeApi);
+        }        
+        
+        let urlNoQuery = url.toLowerCase().split('?')[0];
+
+        if(this._whiteListedPrefixUrls.find(u => urlNoQuery.startsWith(u.toLowerCase()))){
+            return super.send(method, url, body, etag, headers);
+        }
+        else if(urlNoQuery.endsWith(this.tryFunctionApp.site.id.toLowerCase())){
+            return Observable.of(this._getFakeResponse(this.tryFunctionApp.site));
+        }
+        else if(urlNoQuery.endsWith("/providers/microsoft.authorization/permissions")){
+            return Observable.of(this._getFakeResponse({
+                "value" : [{
+                    "actions" : ["*"],
+                    "notActions": []
+                }],
+                "nextLink":null
+            }));
+        }
+        else if(urlNoQuery.endsWith("/providers/microsoft.authorization/locks")){
+            return Observable.of(this._getFakeResponse({"value":[]}));
+        }
+        else if(urlNoQuery.endsWith("/config/web")){
+            return Observable.of(<any>this._getFakeResponse({
+                id : this._tryFunctionApp.site.id,
+                properties : {
+                    scmType : "None"
+                }
+            }));
+        }
+        else if(urlNoQuery.endsWith("/appsettings/list")){
+            return this.tryFunctionApp.getFunctionContainerAppSettings()
+            .map(r =>{
+                return this._getFakeResponse({
+                    properties: r
+                })
+            })
+        }
+
+        this._aiService.trackEvent("/try/arm-send-failure", {
+            uri : url
+        });
+
+        throw "[ArmTryService] - send: " + url;
+    }
+
+    private _getFakeResponse(jsonObj : any) : any{
+        return {
+            headers : {
+                get : (name : string) =>{
+                    return null;
+                }
+            },
+            json : () =>{
+                return jsonObj;
+            }
+        }
     }
 }
